backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
google-cloud-bigquery==3.25.0
python-dotenv==1.0.1
